setup an autoload for this plugin
